functioning on up
will load a empty forum if not config.php is provided

+++ docker/forum_config/forum_setup.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
-cd /app
-if [ -f config.php ]; then
+set -e
+cd "$APPLICATION_PATH"
+
+mkdir -p /app/assets /app/public/assets /app/storage
+chown application -R /app/assets /app/public/assets /app/storage
+
+if [ -f /conf/config.php ]; then
+  ln -s /conf/config.php ./config.php
   php flarum migrate
   # Requires Cache Assets
   # https://discuss.flarum.org/d/23321-cache-assets-by-bokt
@@ -9,7 +15,9 @@ if [ -f config.php ]; then
 
   # Remove old views on load
   # https://discuss.flarum.org/d/25436-beta-14-call-to-undefined-function-array-get/51
-  rm -r storage/views/*
+  if [ -n "$(ls -A storage/views)" ]; then
+    rm -r storage/views/*
+  fi
 else
   cat > install.yml <<EOF
 debug: true
@@ -29,5 +37,7 @@ adminUser:
 settings:
 EOF
   php flarum install --file install.yml
+  cp config.php /conf/config.php
+  chown application /conf/config.php
 fi
